pare au lancement mon general

+++ Sources/Replique.h
@@ -5,5 +5,5 @@ struct Replique
 {
 	std::string text;
 	int scoreMult;
-	int damageMult;
+	float damageMult;
 };
